+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added `Remove-PnPHubToHubAssocation` cmdlet (0.1.9)
 - Added `Remove-PnPSiteUserInvitations` cmdlet (0.1.9)
 - Added `Remove-PnPTenantSyncClientRestriction` cmdlet (0.1.9)
+- Added `Remove-PnPUserInfo` cmdlet (0.1.9)
 
 ### Changed